@@ -1,20 +1,20 @@
-import 'package:animatedbuildersample/src/ui/screen/color.dart';
+import 'package:animated_builder_sample/src/ui/screen/color.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class ScaleAnimationPage extends StatefulWidget {
-  ScaleAnimationPage({Key key}) : super(key: key);
+  const ScaleAnimationPage({super.key});
 
-  static const kRouteName = '/scale';
+  static const kRouteName = 'scale';
 
   @override
-  _ScaleAnimationPageState createState() => _ScaleAnimationPageState();
+  State<ScaleAnimationPage> createState() => _ScaleAnimationPageState();
 }
 
 class _ScaleAnimationPageState extends State<ScaleAnimationPage>
     with SingleTickerProviderStateMixin {
-
-  AnimationController _controller;
+  late AnimationController _controller;
 
   @override
   void initState() {
@@ -37,19 +37,20 @@ class _ScaleAnimationPageState extends State<ScaleAnimationPage>
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('scale animation'),
+        title: const Text('scale animation'),
         actions: <Widget>[
           /// Button to navigate to next page
           CupertinoButton(
-            onPressed: () => Navigator.of(context)
-                .pushNamed(ColorAnimationPage.kRouteName),
+            onPressed: () => context.go(
+                '${GoRouterState.of(context).matchedLocation}/${ColorAnimationPage.kRouteName}'),
             padding: EdgeInsets.zero,
             child: Row(
               children: <Widget>[
-                Text('color',
-                  style: Theme.of(context).primaryTextTheme.headline6,
+                Text(
+                  'color',
+                  style: Theme.of(context).primaryTextTheme.titleLarge,
                 ),
-                Icon(Icons.arrow_forward_ios, color: Colors.white),
+                const Icon(Icons.arrow_forward_ios, color: Colors.white),
               ],
             ),
           ),
@@ -60,7 +61,7 @@ class _ScaleAnimationPageState extends State<ScaleAnimationPage>
         children: <Widget>[
           Row(
             children: <Widget>[
-              Spacer(),
+              const Spacer(),
               CupertinoButton(
                 onPressed: () {
                   if (_controller.status == AnimationStatus.dismissed) {
@@ -70,14 +71,15 @@ class _ScaleAnimationPageState extends State<ScaleAnimationPage>
                   }
                 },
                 color: Colors.green,
-                child: Text('animate',
-                  style: Theme.of(context).primaryTextTheme.headline6,
+                child: Text(
+                  'animate',
+                  style: Theme.of(context).primaryTextTheme.titleLarge,
                 ),
               ),
-              Spacer(),
+              const Spacer(),
             ],
           ),
-          SizedBox(height: 100.0),
+          const SizedBox(height: 100.0),
           WidthAnimation(controller: _controller),
         ],
       ),
@@ -87,22 +89,21 @@ class _ScaleAnimationPageState extends State<ScaleAnimationPage>
 
 class WidthAnimation extends StatelessWidget {
   WidthAnimation({
-    Key key,
-    this.controller,
+    super.key,
+    required this.controller,
   }) : _scale = Tween<double>(begin: 1.0, end: 0.2).animate(
-    /// アニメーションにカーブを設定したいときには
-    /// AnimationControllerをCurvedAnimationで包む
-    CurvedAnimation(
-      parent: controller,
-      curve: Curves.easeInOut,
-    ),
-  ),
-        super(key: key);
+          /// アニメーションにカーブを設定したいときには
+          /// AnimationControllerをCurvedAnimationで包む
+          CurvedAnimation(
+            parent: controller,
+            curve: Curves.easeInOut,
+          ),
+        );
 
   final AnimationController controller;
   final Animation<double> _scale;
 
-  Widget _animationBuilder(BuildContext context, Widget child) {
+  Widget _animationBuilder(BuildContext context, Widget? child) {
     return Transform.scale(
       scale: _scale.value,
       child: child,
@@ -114,6 +115,7 @@ class WidthAnimation extends StatelessWidget {
     return AnimatedBuilder(
       builder: _animationBuilder,
       animation: controller,
+
       /// Transformを使用する場合は、アニメーションがContainer自体のパラメータに
       /// 影響しないのでContainerをまるごとchildに含めて無駄なリビルドを抑制する。
       child: Container(
@@ -121,8 +123,9 @@ class WidthAnimation extends StatelessWidget {
         height: 200.0,
         alignment: Alignment.center,
         color: Colors.red,
-        child: Text('scale',
-          style: Theme.of(context).primaryTextTheme.headline6,
+        child: Text(
+          'scale',
+          style: Theme.of(context).primaryTextTheme.titleLarge,
         ),
       ),
     );
