@@ -1,7 +1,6 @@
-import 'package:animatedbuildersample/src/app.dart';
+import 'package:animated_builder_sample/src/app.dart';
 import 'package:flutter/material.dart';
 
 void main() {
-  runApp(AnimationSampleApp());
+  runApp(const AnimationSampleApp());
 }
-
